ssl_monitor/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
